+++ packages/datadog_flutter_plugin/test/rum/ddrum_method_channel_test.dart
@@ -85,6 +85,14 @@ void main() {
     ]);
   });
 
+  test('addViewLoadingTime calls to platform', () async {
+    await ddRumPlatform.addViewLoadingTime(true);
+
+    expect(log, [
+      isMethodCall('addViewLoadingTime', arguments: {'overwrite': true})
+    ]);
+  });
+
   test('startResource calls to platform', () async {
     await ddRumPlatform.startResource('resource_key', RumHttpMethod.get,
         'https://fakeresource.com/url', {'attribute_key': 'attribute_value'});